fidoip_1.0.5

+++ relinkapps.sh
@@ -16,14 +16,9 @@ f1()
 {
 echo ''
 echo '-----------------------------------------------------'
-echo "Relinking libararies and binaries...                    "
+echo "Relinking libraries and binaries...                    "
 echo '-----------------------------------------------------'
 echo ''
-
-echo "Relinking binkd binary"
-
-cd /usr/local/sbin ; rm -rf binkd
-cd /usr/local/sbin ; ln -sf binkd-0.9.9 binkd
 
 echo "Relinking libraries in /usr/local/lib"
 
@@ -57,22 +52,24 @@ cd /usr/lib64 ; rm -rf libsmapi.so.2.4
 ln -sf /usr/local/lib/libsmapi.so.2.4.0 libsmapi.so.2.4
 
 else
-echo  "Found directory for 32-bit libraries... "
+echo 
+fi
 
+if [ -e /usr/lib ]; 
+then
+echo  "Found directory for 32-bit libraries... "
 cd /usr/lib ; rm -rf libsapi.so
 ln -sf /usr/local/lib/libsmapi.so.2.4.0 libsmapi.so
-
 cd /usr/lib/ ; rm -rf libfidoconfig.so.1.4
 ln -sf /usr/local/lib/libfidoconfig.so.1.4.0 libfidoconfig.so.1.4
-
 cd /usr/lib ; rm -rf libfidoconfig.so
 ln -sf /usr/local/lib/libfidoconfig.so.1.4.0 libfidoconfig.so
-
 cd /usr/lib ; rm -rf libsmapi.so.2
 ln -sf /usr/local/lib/libsmapi.so.2.4 libsmapi.so.2
-
 cd /usr/lib ; rm -rf libsmapi.so.2.4
 ln -sf /usr/local/lib/libsmapi.so.2.4.0 libsmapi.so.2.4
+else
+echo
 fi
 
 echo ''
